% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_token.R
\name{load_token}
\alias{load_token}
\title{Load a dropbox token for use in repeated dropbox calls}
\usage{
load_token()
}
\value{
token object which can be used for passwordless rdrop2 calls
}
\description{
Loads a file that is a dropbox key such that credentials need
not be inputted on every new session. If no key exists yet one is created 
by asking for credentials.
}

