% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_paths.R
\name{model_path}
\alias{model_path}
\title{Load all models from a drop_box designated location}
\usage{
model_path(home = "/ANALÍSIS SIC-MIDO/1 BASES DE DATOS/Margins")
}
\value{
named character vector of file paths
}
\description{
Loads all the dta file paths from a designated location
}

