% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_drop_dta.R
\name{read_drop_dta}
\alias{read_drop_dta}
\title{Read a dropbox dta file}
\usage{
read_drop_dta(file_to_dta)
}
\value{
data frame object read from dropbox dta file path
}
\description{
Read a dropbox dta file by pulling it into a temp file and
using readstata13 to parse it into a data frame
}
