% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf2leaf.R
\name{spdf2leaf}
\alias{spdf2leaf}
\title{render a spatial data frame in leaflet}
\usage{
spdf2leaf(df, col = "data")
}
\value{
leaflet object
}
\description{
Takes a spatial data frame and turns it into an interactive 
leaflet instance.
}

